+++ mnp/index.js
@@ -14,18 +14,41 @@ to create the first page and press enter when done.`)
         }
       },
     },
-    // binary: {
-    //   confirm: true,
-    //   text: 'With binary',
-    //   async afterQuestions() {
+    binary: {
+      confirm: true,
+      text: 'With binary',
+      async afterQuestions({ removeFiles, removePackages, updateFiles }, withBinary ) {
+        if (withBinary) return
+        removeFiles('src/bin')
+        removeFiles('build/bin')
+        removeFiles('test/mask/bin.js')
+        removeFiles('test/result/bin')
+        removeFiles('types/arguments.xml')
+        removeFiles('documentary/2-CLI')
+        await updateFiles({
+          re: /## CLI[\s\S]+#/,
+          replacement: '#',
+        }, { file: 'README.md' })
+        await removePackages(['erte', 'indicatrix', 'usually', 'argufy'])
+      },
+    },
+    // compile: {
+    //   text: 'Build or compile',
+    //   getDefault() { return 'compile' },
+    //   async afterQuestions({}, answer) {
+    //     const compile = answer == 'compile'
+    //     const build = answer == 'build'
+    //     if (compile) {
 
+    //     }
     //   },
     // },
   },
-  async afterInit({ renameFile, name }) {
+  async afterInit({ renameFile, name, initManager }) {
     renameFile('src/bin/mnp.js', `src/bin/${name}.js`)
     renameFile('compile/bin/mnp.js', `compile/bin/${name}.js`)
     renameFile('compile/mnp.js', `compile/${name}.js`)
     renameFile('compile/mnp.js.map', `compile/${name}.js.map`)
+    await initManager()
   },
 }