+++ CHANGELOG.md
@@ -2,5 +2,5 @@
 
 ### 0.0.0-pre
 
-- Create `{{ package-name }}` with _[`My New Package`](https://mnpjs.org)_
+- Create `{{ package-name }}` with _[`My New Package`](https://MNPJS.org)_
 - [repository]: `src`, [`test`](https://contexttesting.com), [`documentary`](https://readme.page) & [`types`](https://typedef.page).